requirements.txt
discord.py[voice]
munch
python-dotenv
youtube-dl
asyncio
pytest-asyncio
pymongo
